Add files via upload
Binary Search Tree (BST)

A BST is a tree-based data structure that efficiently stores and retrieves sorted data. 
Structure: Each node in the BST has a maximum of two child nodes: a left child and a right child.
Ordering Property:
All elements in the left subtree of a node are less than the value of that node.
All elements in the right subtree are greater than the value of that node.